Clarify initial state when creating reconciler
Since the `Usage` example simply a number as the state, the current use of a map made it hard for me to grasp when this map was reused since the rest of the example manipulates the number directly. I propose using nil and a small clarification of the `Controller` section.