Add a single watch to the state per reconciler
Before this patch, each subscription passed to `rum/react` added a watcher
to the state atom of the reconciler through `add-watch`. This means that
every time the atom was `swap!`ed the watch callback was executed, especially
the `(not= oldv newv)`. `not=` can be costly for big/complex enough states.

This patch replaces this mecanism by setting a *single* watch with `add-watch`
on the state per reconciler, at construction time, and registers the watchers
in a simple map. The single watch executes `not=` a single time and executes
the watchers only then, which basically means `rum/request-render`.

For complex enough UI/states it can make a *big* difference in performance.

All tests still pass and no public API has changed.