Don't evaluate default-batched-updates until necessary
The current code uses a `def` meaning that the map will be evaluated at
run time. `js/requestAnimationFrame` and `js/cancelAnimationFrame` only
exist in the browser, so it fails if you run it with Node:

    ReferenceError: requestAnimationFrame is not defined

I used an `or` in order to only evaluate
`(-get-default-batched-updates)` if `batched-updates` is not defined.
This allows one to provide custom batched-updates to run code on Node.

`let`’s `:or` doesn’t work because it *always* evaluates its default
argument:

    (let [{:keys [a]
           :or {a (println "hey")}} {:a 42}]
      a)
    ; "hey"
    ; => 42